requirements.txt
ccxt==3.1.54
pandas==2.0.3
numpy==1.23.5
python-dotenv==1.0.0
supabase==0.7.1
python-telegram-bot==13.15
requests==2.31.0
python-dateutil==2.8.2
google-cloud-secret-manager==2.15.1
psutil==5.9.5
schedule==1.2.1
ntplib==0.4.0
